Recruitment/requirements.txt
APScheduler==3.9.1
arabic-reshaper==2.1.3
asgiref==3.4.1
asn1crypto==1.5.1
asttokens==2.0.5
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.4
comtypes==1.1.11
cryptography==36.0.2
cssselect2==0.6.0
cx-Oracle==8.3.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.1
django-appconf==1.0.5
django-apscheduler==0.6.2
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-cryptography==1.1
django-filter==22.1
django-import-export==2.8.0
django-jazzmin==2.5.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-otp==1.1.3
django-tinymce==3.4.0
django-user-visit==0.5.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
docopt==0.6.2
English-to-Hindi==0.0.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.2
fpdf==1.7.2
future==0.18.2
geographiclib==1.52
geopy==2.2.0
gTTS==2.2.4
html5lib==1.1
idna==3.3
import-export==0.2.67.dev6
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.71
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.9.2
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
MouseInfo==0.1.3
mysqlclient==2.1.0
nbformat==5.4.0
nest-asyncio==1.5.5
numpy==1.22.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
payu-sdk==1.2.0
paywix==1.4.0
pickleshare==0.7.5
Pillow==9.0.0
pipwin==0.5.2
pisa==3.0.33
platformdirs==2.5.1
playsound==1.3.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
PyAudio @ file:///C:/Users/rahul/Downloads/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.12.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyjokes==0.6.0
pyjsparser==2.7.1
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyotp==2.6.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
pySmartDL==1.3.4
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.27.1
requests-oauthlib==1.3.1
schedule==1.1.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
stack-data==0.2.0
svglib==1.4.1
tablib==3.2.1
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
tweepy==4.10.0
tzdata==2021.5
tzlocal==4.2
ua-parser==0.10.0
uritools==4.0.0
urllib3==1.26.8
user-agents==2.2.0
virtualenv==20.13.3
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
whitenoise==6.0.0
wikipedia==1.4.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
